Factor out integ test helper crate to allow running load test separately * Todo: Wft timeout autocomplete needs to be converted to new way of    working now that polling is faster/less predictable
