+++ src/pollers/poll_buffer.rs
@@ -2,58 +2,96 @@ use crate::{
     pollers, protos::temporal::api::workflowservice::v1::PollActivityTaskQueueResponse,
     protos::temporal::api::workflowservice::v1::PollWorkflowTaskQueueResponse, ServerGatewayApis,
 };
-use futures::{stream::repeat_with, Stream, StreamExt};
-use std::future::Future;
-use std::sync::Arc;
-use tokio::sync::Mutex;
+use std::{fmt::Debug, future::Future, sync::Arc};
+use tokio::sync::{
+    mpsc::{channel, Receiver},
+    Mutex, Notify,
+};
 
 pub struct LongPollBuffer<T> {
-    buffered_polls: Mutex<Box<dyn Stream<Item = pollers::Result<T>> + Unpin + Send>>,
+    buffered_polls: Mutex<Receiver<pollers::Result<T>>>,
+    shutdown: Arc<Notify>,
 }
 
 impl<T> LongPollBuffer<T>
 where
-    T: Send,
+    T: Send + Debug + 'static,
 {
-    pub fn new<FT>(poll_fn: impl Fn() -> FT + Send + 'static) -> Self
+    pub fn new<FT>(
+        poll_fn: impl Fn() -> FT + Send + Sync + 'static,
+        concurrent_pollers: usize,
+        buffer_size: usize,
+    ) -> Self
     where
         FT: Future<Output = pollers::Result<T>> + Send,
     {
-        // This is not the world's most efficient thing, but given we're wrapping IO it's unlikely
-        // to be of any significance.
-        let pbuff = repeat_with(poll_fn).buffered(1);
+        let (tx, rx) = channel(buffer_size);
+        let shutdown = Arc::new(Notify::new());
+        let pf = Arc::new(poll_fn);
+        for _ in 0..concurrent_pollers {
+            let tx = tx.clone();
+            let pf = pf.clone();
+            let shutdown = shutdown.clone();
+            tokio::spawn(async move {
+                loop {
+                    let pf = pf();
+                    let r = tokio::select! {
+                        r = pf => r,
+                        _ = shutdown.notified() => break,
+                    };
+                    tx.send(r).await.expect("Long poll buffer is not dropped");
+                }
+            });
+        }
         Self {
-            buffered_polls: Mutex::new(Box::new(pbuff)),
+            buffered_polls: Mutex::new(rx),
+            shutdown,
         }
     }
 
     pub async fn poll(&self) -> pollers::Result<T> {
         let mut locked = self.buffered_polls.lock().await;
         (*locked)
-            .next()
+            .recv()
             .await
             .expect("There is always another item in the stream")
+    }
+
+    pub fn shutdown(&self) {
+        self.shutdown.notify_one();
     }
 }
 
 pub type PollWorkflowTaskBuffer = LongPollBuffer<PollWorkflowTaskQueueResponse>;
 pub fn new_workflow_task_buffer(
     sg: Arc<impl ServerGatewayApis + Send + Sync + 'static>,
+    concurrent_pollers: usize,
+    buffer_size: usize,
 ) -> PollWorkflowTaskBuffer {
-    LongPollBuffer::new(move || {
-        let sg = sg.clone();
-        async move { sg.poll_workflow_task().await }
-    })
+    LongPollBuffer::new(
+        move || {
+            let sg = sg.clone();
+            async move { sg.poll_workflow_task().await }
+        },
+        concurrent_pollers,
+        buffer_size,
+    )
 }
 
 pub type PollActivityTaskBuffer = LongPollBuffer<PollActivityTaskQueueResponse>;
 pub fn new_activity_task_buffer(
     sg: Arc<impl ServerGatewayApis + Send + Sync + 'static>,
+    concurrent_pollers: usize,
+    buffer_size: usize,
 ) -> PollActivityTaskBuffer {
-    LongPollBuffer::new(move || {
-        let sg = sg.clone();
-        async move { sg.poll_activity_task().await }
-    })
+    LongPollBuffer::new(
+        move || {
+            let sg = sg.clone();
+            async move { sg.poll_activity_task().await }
+        },
+        concurrent_pollers,
+        buffer_size,
+    )
 }
 
 #[cfg(test)]
@@ -65,7 +103,7 @@ mod tests {
     use tokio::{select, sync::mpsc::channel};
 
     #[tokio::test]
-    async fn only_polls_once() {
+    async fn only_polls_once_with_1_poller() {
         let mut mock_gateway = MockManualGateway::new();
         mock_gateway
             .expect_poll_workflow_task()
@@ -79,7 +117,7 @@ mod tests {
             });
         let mock_gateway = Arc::new(mock_gateway);
 
-        let pb = new_workflow_task_buffer(mock_gateway);
+        let pb = new_workflow_task_buffer(mock_gateway, 1, 1);
 
         // Poll a bunch of times, "interrupting" it each time, we should only actually have polled
         // once since the poll takes a while
@@ -88,18 +126,21 @@ mod tests {
             interrupter_tx.send(()).await.unwrap();
         }
 
+        // We should never get anything out since we interrupted 100% of polls
         let mut last_val = false;
-        for _ in 0..11 {
+        for _ in 0..10 {
             select! {
                 _ = interrupter_rx.recv() => {
+                    last_val = true;
                 }
                 _ = pb.poll() => {
-                    last_val = true;
                 }
             }
         }
         assert!(last_val);
-        // Now we poll for the second time here, fulfilling our 2 times expectation
+        // Now we grab the buffered poll response, the poll task will go again but we don't grab it,
+        // therefore we will have only polled twice.
         pb.poll().await.unwrap();
+        pb.shutdown();
     }
 }
