+++ tests/integ_tests/heartbeat_tests.rs
@@ -1,5 +1,5 @@
-use crate::integ_tests::{init_core_and_create_wf, schedule_activity_cmd};
 use assert_matches::assert_matches;
+use integ_test_helpers::{init_core_and_create_wf, schedule_activity_cmd};
 use std::time::Duration;
 use temporal_sdk_core::protos::coresdk::{
     activity_result::{self, activity_result as act_res, ActivityResult},
